moved file to ConditionalAlgorithm
