+++ src/ConditionalAlgorithm.ts
